Atualizando a configuração no banco de dados

+++ src/app/data_classes/cliente.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 
 @dataclass
 class Cliente:
@@ -9,6 +10,6 @@ class Cliente:
     nome: str
     data_nasc: date
     email: str
-    salario: float
-    vl_patrimonio: float
+    salario: Decimal
+    vl_patrimonio: Decimal
     id_profissao: int
